Move termHeader into a new component
