+++ js/components/term/index.jsx
@@ -3,11 +3,10 @@ import { connect } from 'react-redux';
 
 // Internal dependencies
 import QueryPosts from 'components/data/query-posts';
-import QueryTerm from 'components/data/query-term';
 import { isRequestingPostsForQuery, getPostsForQuery, getTotalPagesForQuery } from 'state/posts/selectors';
-import { isRequestingTerm, getTermIdFromSlug, getTerm } from 'state/terms/selectors';
 
 // Components
+import TermHeader from './header';
 import PostList from '../posts/list';
 import Pagination from '../pagination/archive';
 
@@ -22,18 +21,10 @@ const Term = React.createClass( {
 
 	render() {
 		const posts = this.props.posts || [];
-		const termData = this.props.termData || {};
 
 		return (
 			<div className="card">
-				<QueryTerm taxonomy={ this.props.taxonomy } termSlug={ this.props.term } />
-				{ this.props.requestingTerm ?
-					this.renderHeaderPlaceholder() :
-					<header className="page-header">
-						<h1 className="page-title">{ termData.name }</h1>
-						{ termData.description && <p>{ termData.description }</p> }
-					</header>
-				}
+				<TermHeader params={ { taxonomy: this.props.route.taxonomy, slug: this.props.params.slug } } />
 
 				<QueryPosts query={ this.props.query } />
 				{ this.props.requesting ?
@@ -56,19 +47,11 @@ export default connect( ( state, ownProps ) => {
 		query.tag = ownProps.params.slug;
 	}
 
-	const term = ownProps.params.slug;
-	const taxonomy = ownProps.route.taxonomy;
-	const termId = getTermIdFromSlug( state, taxonomy, term );
-
 	return {
-		term,
-		taxonomy,
 		query,
-		termData: getTerm( state, termId ),
 		page: parseInt( query.paged ),
 		posts: getPostsForQuery( state, query ),
 		totalPages: getTotalPagesForQuery( state, query ),
 		requesting: isRequestingPostsForQuery( state, query ),
-		requestingTerm: isRequestingTerm( state, taxonomy, term )
 	};
 } )( Term );
